UBI-315 Fixed mute button toggle state inconsistencies after disconnecting and reconnecting to the Persona
- Objective-C sample:
Set the `isMuted` flag to false on disconnection. This fixes an issue where the mute button does not toggle on on the first touch after disconnecting (with the mute button toggled on) and reconnecting to the Persona.

- Swift sample:
Set the `isMuted` flag to false and the mute button icon image to "mic.fill" on disconnection. This fixes an issue where we attempt to mute the Persona when it is already muted (resulting in an error and therefore failing to set the correct "mic.fill" mute button icon image) after disconnecting (with the mute button toggled on) and reconnecting to the Persona.

Risk areas: Mute button for the Objective-C and Swift sample. A smoke test is recommended